removr color.c
